+++ src/index.ts
@@ -1,6 +1,8 @@
 // export * from './deck-libs';
 export * from './custom-components';
+export * from './custom-hooks';
 export * from './deck-components';
+export * from './deck-hooks'
 export * from './plugin';
 export * from './webpack';
 export * from './utils';